# hack/contrib/docker/api/html/swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/v2/tenants/query/{tenant_name}": {
      "get": {
        "description": "get tenant resources",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "租户带资源列表",
        "operationId": "tenants",
        "parameters": [
          {
            "type": "string",
            "format": "string",
            "description": "123",
            "name": "tenant_name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "统一返回格式",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
...
    "/v2/tenants/{tenant_name}/plugin": {
      "get": {
        "description": "get plugins",
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "获取当前租户下所有的可用插件",
        "operationId": "getPlugins",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "name": "tenant_name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "统一返回格式",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      },
      "post": {
        "description": "create plugin",
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "创建插件",
        "operationId": "createPlugin",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "name": "tenant_name",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "plugin_id",
                "plugin_name"
              ],
              "properties": {
                "build_model": {
                  "description": "构建模式\nin: body",
                  "type": "string",
                  "x-go-name": "BuildModel"
                },
                "git_url": {
                  "description": "git地址\nin: body",
                  "type": "string",
                  "x-go-name": "GitURL"
                },
                "image_url": {
                  "description": "插件docker地址\nin:body",
                  "type": "string",
                  "x-go-name": "ImageURL"
                },
                "plugin_id": {
                  "description": "插件id\nin: body",
                  "type": "string",
                  "x-go-name": "PluginID"
                },
                "plugin_info": {
                  "description": "插件用途描述\nin: body",
                  "type": "string",
                  "x-go-name": "PluginInfo"
                },
                "plugin_model": {
                  "description": "插件模式\nin: body",
                  "type": "string",
                  "x-go-name": "PluginModel"
                },
                "plugin_name": {
                  "description": "in: body",
                  "type": "string",
                  "x-go-name": "PluginName"
                },
                "tenant_id": {
                  "description": "租户id\nin: body",
                  "type": "string",
                  "x-go-name": "TenantID"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": "统一返回格式",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
    "/v2/tenants/{tenant_name}/plugin/{plugin_id}": {
      "put": {
        "description": "update plugin",
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "插件更新 全量更新，但pluginID和所在租户不提供修改",
        "operationId": "updatePlugin",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "description": "租户名称",
            "name": "tenant_name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "PluginID",
            "description": "插件id",
            "name": "plugin_id",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "build_model": {
                  "description": "构建模式\nin: body",
                  "type": "string",
                  "x-go-name": "BuildModel"
                },
                "git_url": {
                  "description": "git地址\nin: body",
                  "type": "string",
                  "x-go-name": "GitURL"
                },
                "image_url": {
                  "description": "插件docker地址\nin: body",
                  "type": "string",
                  "x-go-name": "ImageURL"
                },
                "plugin_info": {
                  "description": "插件用途描述\nin: body",
                  "type": "string",
                  "x-go-name": "PluginInfo"
                },
                "plugin_model": {
                  "description": "插件模式\nin: body",
                  "type": "string",
                  "x-go-name": "PluginModel"
                },
                "plugin_name": {
                  "description": "插件名称\nin: body",
                  "type": "string",
                  "x-go-name": "PluginName"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": "统一返回格式",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      },
      "delete": {
        "description": "delete plugin",
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "插件删除",
        "operationId": "deletePlugin",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "name": "tenant_name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "PluginID",
            "name": "plugin_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "统一返回格式",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
    "/v2/tenants/{tenant_name}/plugin/{plugin_id}/build": {
      "post": {
        "description": "build plugin",
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "构建plugin",
        "operationId": "buildPlugin",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "name": "tenant_name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "PluginID",
            "name": "plugin_id",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "plugin_cpu",
                "plugin_memory",
                "build_version",
                "info",
                "tenant_id"
              ],
              "properties": {
                "build_image": {
                  "description": "镜像地址\nin: body",
                  "type": "string",
                  "x-go-name": "BuildImage"
                },
                "build_version": {
                  "description": "部署的版本号\nin: body",
                  "type": "string",
                  "x-go-name": "BuildVersion"
                },
                "event_id": {
                  "description": "the event id\nin: body",
                  "type": "string",
                  "x-go-name": "EventID"
                },
                "info": {
                  "description": "版本信息, 协助选择插件版本\nin:body",
                  "type": "string",
                  "x-go-name": "Info"
                },
                "operator": {
                  "description": "操作人\nin: body",
                  "type": "string",
                  "x-go-name": "Operator"
                },
                "plugin_cmd": {
                  "description": "插件cmd, 默认50\nin: body",
                  "type": "string",
                  "x-go-name": "PluginCMD"
                },
                "plugin_cpu": {
                  "description": "插件CPU权重, 默认125\nin: body",
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "PluginCPU"
                },
                "plugin_from": {
                  "description": "安装来源\nin: body",
                  "type": "string",
                  "x-go-name": "PluginFrom"
                },
                "plugin_memory": {
                  "description": "插件最大内存, 默认50\nin: body",
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "PluginMemory"
                },
                "repo_url": {
                  "description": "git地址分支信息，默认为master\nin: body",
                  "type": "string",
                  "x-go-name": "RepoURL"
                },
                "tenant_id": {
                  "description": "租户id\nin: body",
                  "type": "string",
                  "x-go-name": "TenantID"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": "统一返回格式",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
...
    "/v2/tenants/{tenant_name}/plugin/{plugin_id}/build-version/{version_id}": {
      "get": {
        "description": "plugin version",
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "获取某个构建版本信息",
        "operationId": "pluginVersion",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "name": "tenant_name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "PluginID",
            "name": "plugin_id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "VersionID",
            "name": "version_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "统一返回格式",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      },
      "delete": {
        "description": "delete plugin version",
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "删除某个构建版本信息",
        "operationId": "deletePluginVersion",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "name": "tenant_name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "PluginID",
            "name": "plugin_id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "VersionID",
            "name": "version_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "统一返回格式",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
...
    "/v2/tenants/{tenant_name}/res": {
      "get": {
        "description": "get tenant resources",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "租户带资源单个",
        "operationId": "tenants",
        "parameters": [
          {
            "type": "string",
            "format": "string",
            "description": "123",
            "name": "tenant_name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "统一返回格式",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
...
    "/v2/tenants/{tenant_name}/services/{service_alias}/plugin/{plugin_id}": {
      "delete": {
        "description": "delete plugin relation",
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "删除插件依赖",
        "operationId": "deletePluginRelation",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "name": "tenant_name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ServiceAlias",
            "name": "service_alias",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "PluginID",
            "description": "插件id",
            "name": "plugin_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "统一返回格式",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
    "/v2/tenants/{tenant_name}/services/{service_alias}/plugin/{plugin_id}/envs": {
      "get": {
        "description": "get version env",
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "v2"
        ],
        "summary": "获取可配置的env; 从service plugin对应中取, 若不存在则返回默认可修改的变量",
        "operationId": "getVersionEnvs",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TenantName",
            "name": "tenant_name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ServiceAlias",
            "name": "service_alias",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "PluginID",
            "description": "插件id",
            "name": "plugin_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "统一返回格式",
            "schema": {
              "$ref": "#/responses/commandResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Values": {
      "description": "It is typically used for query parameters and form values.\nUnlike in the http.Header map, the keys in a Values map\nare case-sensitive.",
      "type": "object",
      "title": "Values maps a string key to a list of values.",
      "additionalProperties": {
        "type": "array",
        "items": {
          "type": "string"
        }
      },
      "x-go-package": "net/url"
    },
...
  },
  "responses": {
    "commandResponse": {
      "description": "CommandResponse api统一返回结构",
      "schema": {
        "type": "object",
        "properties": {
          "bean": {
            "description": "单资源实体",
            "type": "object",
            "x-go-name": "Bean"
          },
          "list": {
            "description": "资源列表",
            "type": "object",
            "x-go-name": "List"
          },
          "msg": {
            "description": "API错误信息",
            "type": "string",
            "x-go-name": "Msg"
          },
          "number": {
            "description": "数据集总数",
            "type": "integer",
            "format": "int64",
            "x-go-name": "ListAllNumber"
          },
          "page": {
            "description": "当前页码数",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Page"
          },
          "validation_error": {
            "$ref": "#/definitions/Values"
          }
        }
      }
    }
  }
}